refactor constructor
