+++ src/bytecode/todo.md
@@ -1,6 +1,6 @@
 # TODO
 
-- Finish Runtime implementation.
 - Default signatures for types and runtime. Also export them so that they can be extended.
+- Runtime tests.
 - Polish implementation.
 - Docs.
