--- src/bytecode/types/mask.ts
@@ -1,77 +1,0 @@
-export type UINT8_ZERO_MASK =
-  | '00000000000000000000000000000000000000000000000000000000000000'
-  | '000000000000000000000000000000000000000000000000000000000000000'
-export type UINT16_ZERO_MASK =
-  | '000000000000000000000000000000000000000000000000000000000000'
-  | '0000000000000000000000000000000000000000000000000000000000000'
-export type UINT24_ZERO_MASK =
-  | '000000000000000000000000000000000000000000000000000000000'
-  | '0000000000000000000000000000000000000000000000000000000000'
-export type UINT32_ZERO_MASK =
-  | '00000000000000000000000000000000000000000000000000000000'
-  | '000000000000000000000000000000000000000000000000000000000'
-export type UINT40_ZERO_MASK =
-  | '000000000000000000000000000000000000000000000000000000'
-  | '0000000000000000000000000000000000000000000000000000000'
-export type UINT48_ZERO_MASK =
-  | '0000000000000000000000000000000000000000000000000000'
-  | '00000000000000000000000000000000000000000000000000000'
-export type UINT56_ZERO_MASK =
-  | '00000000000000000000000000000000000000000000000000'
-  | '000000000000000000000000000000000000000000000000000'
-export type UINT64_ZERO_MASK =
-  | '000000000000000000000000000000000000000000000000'
-  | '0000000000000000000000000000000000000000000000000'
-export type UINT72_ZERO_MASK =
-  | '0000000000000000000000000000000000000000000000'
-  | '00000000000000000000000000000000000000000000000'
-export type UINT80_ZERO_MASK =
-  | '00000000000000000000000000000000000000000000'
-  | '000000000000000000000000000000000000000000000'
-export type UINT88_ZERO_MASK =
-  | '000000000000000000000000000000000000000000'
-  | '0000000000000000000000000000000000000000000'
-export type UINT96_ZERO_MASK =
-  | '0000000000000000000000000000000000000000'
-  | '00000000000000000000000000000000000000000'
-export type UINT104_ZERO_MASK =
-  | '00000000000000000000000000000000000000'
-  | '000000000000000000000000000000000000000'
-export type UINT112_ZERO_MASK =
-  | '000000000000000000000000000000000000'
-  | '0000000000000000000000000000000000000'
-export type UINT120_ZERO_MASK =
-  | '0000000000000000000000000000000000'
-  | '00000000000000000000000000000000000'
-export type UINT128_ZERO_MASK =
-  | '00000000000000000000000000000000'
-  | '000000000000000000000000000000000'
-export type UINT136_ZERO_MASK =
-  | '000000000000000000000000000000'
-  | '0000000000000000000000000000000'
-export type UINT144_ZERO_MASK =
-  | '0000000000000000000000000000'
-  | '00000000000000000000000000000'
-export type UINT152_ZERO_MASK =
-  | '00000000000000000000000000'
-  | '000000000000000000000000000'
-export type ADDRESS_ZERO_MASK =
-  | '000000000000000000000000'
-  | '0000000000000000000000000'
-export type UINT168_ZERO_MASK =
-  | '0000000000000000000000'
-  | '00000000000000000000000'
-export type UINT176_ZERO_MASK = '00000000000000000000' | '000000000000000000000'
-export type UINT184_ZERO_MASK = '000000000000000000' | '0000000000000000000'
-export type UINT192_ZERO_MASK = '0000000000000000' | '00000000000000000'
-export type UINT200_ZERO_MASK = '00000000000000' | '000000000000000'
-export type UINT208_ZERO_MASK = '000000000000' | '0000000000000'
-export type UINT216_ZERO_MASK = '0000000000' | '00000000000'
-export type UINT224_ZERO_MASK = '00000000' | '000000000'
-export type UINT232_ZERO_MASK = '000000' | '0000000'
-export type UINT240_ZERO_MASK = '0000' | '00000'
-export type UINT248_ZERO_MASK = '00' | '000'
-export type RevertErrorString =
-  '08c379a000000000000000000000000000000000000000000000000000000000'
-export type RevertPanicString =
-  '4e487b7100000000000000000000000000000000000000000000000000000000'
