+++ src/bytecode/types/opcodes.ts
@@ -216,3 +216,9 @@ export type DUPN =
   | OPCODES['DUP14']
   | OPCODES['DUP15']
   | OPCODES['DUP16']
+
+export type RevertErrorString =
+  '08c379a000000000000000000000000000000000000000000000000000000000'
+
+export type RevertPanicString =
+  '4e487b7100000000000000000000000000000000000000000000000000000000'
